refactor(clerk-js): Refactor how canActivate is used in Routes
fix(clerk-js): Correct useNavigateToFlowStart imports
