+++ packages/clerk-js/src/ui/hooks/index.ts
@@ -17,3 +17,4 @@ export * from './useSetTimeout';
 export * from './useScrollLock';
 export * from './useDeepEqualMemo';
 export * from './useClerkModalStateParams';
+export * from './useNavigateToFlowStart';
