+++ packages/clerk-js/src/ui/router/HashRouter.tsx
@@ -41,7 +41,6 @@ export const HashRouter = ({ preservedParams, children }: HashRouterProps): JSX.
     <BaseRouter
       getPath={getPath}
       basePath={hashRouterBase}
-      flowStartPath={''}
       startPath={''}
       getQueryString={getQueryString}
       internalNavigate={internalNavigate}
