+++ packages/clerk-js/src/ui/router/PathRouter.tsx
@@ -53,7 +53,6 @@ export const PathRouter = ({ basePath, preservedParams, children }: PathRouterPr
     <BaseRouter
       basePath={basePath.substring(1)}
       startPath={''}
-      flowStartPath={''}
       getPath={getPath}
       getQueryString={getQueryString}
       internalNavigate={internalNavigate}
